+++ tests/test_upload.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from src.image_upload_cli.upload import UPLOAD, imgur_upload
+from src.images_upload_cli.upload import UPLOAD, imgur_upload
 
 
 def test_get_upload_func():
